test: export 3d png, failed yet

+++ src/utils/FontUtil.cpp
@@ -4,13 +4,13 @@
 
 #include "FontUtil.h"
 
-int FontUtil::getTextFontWidth(const QString& data, const int fontSize, const float letterSpacing)
+int FontUtil::getTextFontWidth(const QString& data, const double fontSize, const float letterSpacing)
 {
     const int dataLength = data.length();
     return qCeil(dataLength * SIM_PER_POUND_WIDTH * fontSize + (dataLength - 1) * SIM_PER_POUND_WIDTH * letterSpacing);
 }
 
-int FontUtil::getSingleTextSize(const float fontSize)
+int FontUtil::getSingleTextSize(const double fontSize)
 {
     return qCeil(SIM_PER_POUND_WIDTH * fontSize);
 }
